add edit functions

+++ Data/Models/CreditCard.cs
@@ -9,7 +9,7 @@ namespace Data.Models
     public class CreditCard
     {
         public int Id { get; set; }
-        public int Number { get; set; }
+        public Int64 Number { get; set; }
         public int CVV { get; set; }
         public int Year_ExpirationDate { get; set; }
         public int Month_ExpirationDate { get; set; }
